backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Django REST Framework backend for the schematic mapper application"
requires-python = ">=3.12"
dependencies = [
    "cairosvg>=2.8.2",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
    "django-auth-ldap>=4.7.0",
    "ldap3>=2.9",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.9",
    "djangorestframework-stubs>=3.16.8",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.5",
    "types-ldap3",
]
